[percentile] distribution code reorg
* Move distribution.go and context_sketch.go to the metrics package
* Move some helper functions to test_helper.go
* Fix gk_array test
* Remove redundant test
