+++ pkg/metrics/distribution_test.go
@@ -3,14 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-package aggregator
+package metrics
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/DataDog/datadog-agent/pkg/metrics"
 	"github.com/DataDog/datadog-agent/pkg/metrics/percentile"
 )
 
@@ -23,9 +22,9 @@ func TestDistributionSampling(t *testing.T) {
 
 	// Add metric samples and check that the flushed summary series
 	// are correct
-	distro.addSample(&metrics.MetricSample{Value: 1}, 10)
-	distro.addSample(&metrics.MetricSample{Value: 10}, 11)
-	distro.addSample(&metrics.MetricSample{Value: 5}, 12)
+	distro.addSample(&MetricSample{Value: 1}, 10)
+	distro.addSample(&MetricSample{Value: 10}, 11)
+	distro.addSample(&MetricSample{Value: 5}, 12)
 
 	assert.Equal(t, int64(3), distro.count)
